Clean code after review - extract methods and abstract class
